Refactor autocomplete
