+++ src/components/address-autocomplete-field.scss
@@ -1,0 +1,44 @@
+// Shared by venue-detail block and venue navigator (Popover + suggestion list).
+.gatherpress-venue-detail__address-popover .components-popover__content {
+	width: min(100vw - 32px, 28rem);
+	padding: 0;
+	overflow: hidden;
+}
+
+.gatherpress-venue-detail__address-popover-inner {
+	display: flex;
+	align-items: center;
+	gap: 8px;
+	padding: 12px 16px;
+}
+
+.gatherpress-venue-detail__address-loading {
+	font-size: 13px;
+}
+
+.gatherpress-venue-detail__address-suggestions {
+	margin: 0;
+	padding: 4px 0;
+	list-style: none;
+	max-height: 220px;
+	overflow-y: auto;
+}
+
+.gatherpress-venue-detail__address-suggestion {
+	display: block;
+	width: 100%;
+	margin: 0;
+	padding: 8px 12px;
+	text-align: left;
+	border: none;
+	background: transparent;
+	font: inherit;
+	line-height: 1.4;
+	cursor: pointer;
+	color: #1e1e1e;
+}
+
+.gatherpress-venue-detail__address-suggestion:hover,
+.gatherpress-venue-detail__address-suggestion.is-active {
+	background: #f0f0f0;
+}
